Shift function changes and move generation
Shift function changes - Centralised the shift functions so that now instead of having one for each direction, there is just one which takes the list of directions stored in an array and the chosen bitboard and does the shift based on the chosen direction, along with applying a mask.

Move generation - another step for move generation has been completed,  using the Dump7Fill algorithm it looks at each direction of a piece: horizontally, vertically and diagonally and looks at where the next free space is between it and all the opponent pieces in the way. It then adds to a legalMoves variable and then returns that variable once all moves have been found.

printSingleBitboard function - added a function to better see single bitboards with either one or zero bits printed  so that i can use it for other bitboards than just black or white pieces.

Small changes - changed the macro names to SCREAMING_SNAKE_CASE standard, added some comments, changed the squares enum format to correctly print the bitboard