Function to check if move is valid and commit moves
isValid function - checks if a move is valid for a bitboard and the coordinates of the square

commitMove function - using a similar method to generate moves, it finds all disks that can be captured in all directions between the new disk and any other disks of the same colour. It then correctly updates each bitboard to the changes.